Clean Code

+++ stream/src/stream/TripDauration.java
@@ -1,6 +1,10 @@
 package stream;
 
+/**
+ * @author Ahmednasser
+ *
+ */
 public class TripDauration {
-int totalTrips=0;
-long totalTime=0;
+	int totalTrips = 0;
+	Double totalTime = 0.0;
 }
